add token to Tags

+++ src/main/java/tags/TagHandler.java
@@ -4,51 +4,54 @@ import com.sun.net.httpserver.Headers;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import tools.HttpMethods;
+import user.UserQueries;
 
 import java.io.IOException;
 import java.io.OutputStream;
 
 public class TagHandler implements HttpHandler {
 
-    public static final String TAG = "/tag"; // -> GET "/user"
+    public static final String TAG = "/tag";
     public static final String TAGS = "/tags";
     public static final String RISKSFORTAG = "/tag/risks";
 
     public void handle(HttpExchange t) throws IOException {
-        Headers tmp = t.getRequestHeaders();
+        Headers headers = t.getRequestHeaders();
         String response = "";
-        String tagName;
-        String riskId;
+        String tagName = headers.getFirst("TagName");
+        String riskId = headers.getFirst("riskId");
+
+        String token = headers.getFirst("X-Token");
+        String method = t.getRequestMethod();
+
         try {
+            if (!UserQueries.checkAuthorization(token))
+                throw new Exception("Unauthorized");
+
             switch (t.getRequestURI().toString()) {
                 case TAG:
-                    String method = t.getRequestMethod();
                     switch (method) {
                         case HttpMethods.PUT:
-                            tagName = tmp.getFirst("TagName");
-                            riskId = tmp.getFirst("riskId");
                             response = TagQueries.createTag(tagName, riskId);
                             break;
                         case HttpMethods.DELETE:
-                            tagName = tmp.getFirst("TagName");
-                            riskId = tmp.getFirst("riskId");
                             response = TagQueries.deleteTag(tagName, riskId);
                             break;
                     }
                     break;
                 case TAGS:
-                    riskId = tmp.getFirst("riskId");
                     response = TagQueries.getTags(riskId);
                     break;
                 case RISKSFORTAG:
-                    tagName = tmp.getFirst("TagName");
                     response = TagQueries.getRisks(tagName);
                     break;
                 default:
                     throw new NoSuchMethodException();
             }
+
             if (response.contains("ERROR"))
                 throw new Exception(response);
+
             t.sendResponseHeaders(200, response.length());
             OutputStream os = t.getResponseBody();
             os.write(response.getBytes());
@@ -56,6 +59,5 @@ public class TagHandler implements HttpHandler {
         } catch (Exception ex) {
             t.sendResponseHeaders(500, 0);
         }
-
     }
 }
